csv_parser/requirements.txt
pandas
openpyxl
pyyaml
jsonpath-ng
